Shift project whitelist logic to jobservice
