+++ python/feast_spark/job_service.py
@@ -5,7 +5,7 @@ import threading
 import time
 import traceback
 from concurrent.futures import ThreadPoolExecutor
-from typing import Dict, List, Tuple, cast
+from typing import Dict, List, Optional, Tuple, cast
 
 import grpc
 from google.api_core.exceptions import FailedPrecondition
@@ -98,12 +98,25 @@ class JobServiceServicer(JobService_pb2_grpc.JobServiceServicer):
     def __init__(self, client: Client):
         self.client = client
 
+    @property
+    def _whitelisted_projects(self) -> Optional[List[str]]:
+        if self.client.config.exists(opt.WHITELISTED_PROJECTS):
+            whitelisted_projects = self.client.config.get(opt.WHITELISTED_PROJECTS)
+            return whitelisted_projects.split(",")
+        return None
+
+    def is_whitelisted(self, project: str):
+        # Whitelisted projects not specified, allow all projects
+        if not self._whitelisted_projects:
+            return True
+        return project in self._whitelisted_projects
+
     def StartOfflineToOnlineIngestionJob(
         self, request: StartOfflineToOnlineIngestionJobRequest, context
     ):
         """Start job to ingest data from offline store into online store"""
 
-        if not self.client.is_whitelisted(request.project):
+        if not self.is_whitelisted(request.project):
             raise ValueError(
                 f"Project {request.project} is not whitelisted. Please contact your Feast administrator to whitelist it."
             )
@@ -132,7 +145,7 @@ class JobServiceServicer(JobService_pb2_grpc.JobServiceServicer):
     def GetHistoricalFeatures(self, request: GetHistoricalFeaturesRequest, context):
         """Produce a training dataset, return a job id that will provide a file reference"""
 
-        if not self.client.is_whitelisted(request.project):
+        if not self.is_whitelisted(request.project):
             raise ValueError(
                 f"Project {request.project} is not whitelisted. Please contact your Feast administrator to whitelist it."
             )
@@ -164,7 +177,7 @@ class JobServiceServicer(JobService_pb2_grpc.JobServiceServicer):
     ):
         """Start job to ingest data from stream into online store"""
 
-        if not self.client.is_whitelisted(request.project):
+        if not self.is_whitelisted(request.project):
             raise ValueError(
                 f"Project {request.project} is not whitelisted. Please contact your Feast administrator to whitelist it."
             )
@@ -214,7 +227,7 @@ class JobServiceServicer(JobService_pb2_grpc.JobServiceServicer):
     def ListJobs(self, request, context):
         """List all types of jobs"""
 
-        if not self.client.is_whitelisted(request.project):
+        if not self.is_whitelisted(request.project):
             raise ValueError(
                 f"Project {request.project} is not whitelisted. Please contact your Feast administrator to whitelist it."
             )
@@ -349,9 +362,11 @@ def ensure_stream_ingestion_jobs(client: Client, all_projects: bool):
         if all_projects
         else [client.feature_store.project]
     )
-    if client._whitelisted_projects:
+    if client._job_service._whitelisted_projects:
         projects = [
-            project for project in projects if project in client._whitelisted_projects
+            project
+            for project in projects
+            if project in client._job_service._whitelisted_projects
         ]
 
     expected_job_hash_to_tables = _get_expected_job_hash_to_tables(client, projects)
